+++ src/main/java/com/ols/ols_project/mapper/UserMapper.java
@@ -13,18 +13,18 @@ import java.util.List;
  */
 public interface UserMapper {
 
-    UserEntity getUserInfoById(int id);
+    UserEntity getUserInfoById(long id);
 
-    String getPassWodById(int id);
+    String getPassWodById(long id);
 
-    int changePassWordById(int id,String passWord);
+    int changePassWordById(long id,String passWord);
 
-    List<List<AcceptTask>> getAcceptTaskByUserId(int id, String query, int start, int end, String queryInfo, String searchInfo);
+    List<List<AcceptTask>> getAcceptTaskByUserId(long id, String query, int start, int end, String queryInfo, String searchInfo);
 
-    List<List<TaskEntity>> getReleaseTaskByUserId(int id, String query, int start, int end, String queryInfo, String searchInfo);
+    List<List<TaskEntity>> getReleaseTaskByUserId(long id, String query, int start, int end, String queryInfo, String searchInfo);
 
     List<List<UserSignUp>> getReviewerSignUp(String queryInfo, String searchInfo , int start, Integer end);
 
-    int yesAndNoReviewerSignUp(int userId,String operation);
+    int yesAndNoReviewerSignUp(long userId,String operation);
 
 }
