+++ src/main/java/com/ols/ols_project/model/AcceptTaskBo.java
@@ -19,7 +19,7 @@ public class AcceptTaskBo {
     /**
      * 任务编号
      */
-    private Integer taskId;
+    private Long taskId;
 
     /**
      * 任务名称
