+++ src/main/java/com/ols/ols_project/model/FinishCheckTaskBo.java
@@ -20,7 +20,7 @@ public class FinishCheckTaskBo {
     /**
      * 任务ID
      */
-    private Integer taskId;
+    private Long taskId;
 
     /**
      * 任务名称
@@ -40,7 +40,7 @@ public class FinishCheckTaskBo {
     /**
      * 审核编号
      */
-    private Integer judgeId;
+    private Long judgeId;
 
     /**
      * 是否通过审核
