+++ src/main/java/com/ols/ols_project/model/TaskEntityBo.java
@@ -23,7 +23,7 @@ public class TaskEntityBo {
      * 任务编号
      * 主键
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 名称
@@ -68,7 +68,7 @@ public class TaskEntityBo {
     /**
      *发布者编号
      */
-    private Integer release_user_id;
+    private Long release_user_id;
 
     /**
      *接受者数量
@@ -78,7 +78,7 @@ public class TaskEntityBo {
     /**
      *采纳的接受任务编号
      */
-    private Integer adopt_accept_id;
+    private Long adopt_accept_id;
 
     /**
      * 备用字段1
